refactor(open-mpm): split search/ oversized files (closes #365)
Split two oversized files in crates/open-mpm/src/search/ under the
500-line cap, pure reorganization with no logic changes:

- indexer.rs (2163) -> indexer/ {mod, chunker, index, search, tests/*}
- service.rs (745)  -> service/ {mod, query, tests}

The public interface (search::indexer::{CodeChunk, CodeIndexer,
DEFAULT_COOL_AFTER_MINUTES}; search::service::{run_search_service,
is_daemon_running, ...}) is unchanged; thin mod.rs files re-export so
all external crate::search::{indexer,service}::* paths resolve as before.

Part of #356
